tests/wpt/web-platform-tests/service-workers/service-worker/activation-after-registration.https.html

<!DOCTYPE html>
<title>Service Worker: Activation occurs after registration</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/test-helpers.sub.js"></script>
<body>
<script>
var t = async_test('activation occurs after registration');
t.step(function() {
    var scope = 'resources/blank.html';
    var registration;

    service_worker_unregister_and_register(
        t, 'resources/empty-worker.js', scope)
      .then(function(r) {
          registration = r;
          assert_equals(
              r.installing.state,
              'installing',
              'worker should be in the "installing" state upon registration');
          return wait_for_state(t, r.installing, 'activated');
        })
      .then(function() {
          service_worker_unregister_and_done(t, scope);
        })
      .catch(unreached_rejection(t));
});
</script>
</body>
